requirements.txt

---

### `/opt/trovix-capi/requirements.txt`
```txt
Flask==3.0.0
gunicorn==21.2.0
redis==5.0.1
rq==1.15.1
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
structlog==24.1.0
sqlalchemy==2.0.36
PyYAML==6.0.2
pytest==8.3.2
# Optional extras
uvicorn==0.30.1         # alt ASGI server if you ever go FastAPI
black==24.4.2           # code formatter
